+++ App.js
@@ -2,30 +2,18 @@ import React from 'react';
 import {Image, StyleSheet, View} from 'react-native';
 import InputTester from "./components/InputTester";
 import ButtonTester from "./components/ButtonTester";
+import {createStackNavigator} from "react-navigation";
+import WelcomeComponent from "./components/WelcomeComponent";
 
-export default class App extends React.Component {
-    render() {
-        let pic = {
-            uri: 'https://upload.wikimedia.org/wikipedia/commons/d/de/Bananavarieties.jpg'
-        }
-        return (
-            <View style={styles.container}>
-                <Image source={pic} style={{
-                    width: 193,
-                    height: 110
-                }} />
-                <InputTester/>
-                <ButtonTester/>
-            </View>
-        );
-    }
-}
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#fff',
-        alignItems: 'center',
-        justifyContent: 'center',
+const App = createStackNavigator(
+    {
+        Welcome: {
+            screen: WelcomeComponent,
+            navigationOptions: {
+                header: null
+            }
+        },
+        InputTester_: { screen: InputTester }
     },
-});
+);
+export default App
