+++ components/ButtonTester.js
@@ -5,16 +5,24 @@ export default class ButtonTester extends Component {
   constructor(props) {
     super(props);
   }
+  static navigationOptions = {
+    title: 'Welcome',
+  };
 
   render() {
-    return (
-      <View style={{padding: 10}}>
-          <Button onPress={() => {
-              Alert.alert('You tapped the button')
-          }}
-                  title="Tap this"
-          />
-      </View>
-    );
+      const { navigate } = this.props.navigation;
+      return (
+          <View style={{padding: 10}}>
+              <Button onPress={() => {
+                  Alert.alert('You tapped the button')
+              }}
+                      title="Tap this"
+              />
+              <Button
+                  title="Go to Jane's profile"
+                  onPress={() => navigate('InputTester_')}
+              />
+          </View>
+      );
   }
 }
